Refactor organization
Astro was complaining about "png" and ".svelte" not being supported
extensions for pages. Which is correct - images should go on the public
folder, and pages are just components, loaded by a .astro file. So this is
fixed with this commit. This will also fix the images not being loaded in
the live website.